test: 리팩토링에 따른 test 수정
